Rename TcnBluetoothService to TcnBluetoothManager
